+++ archetypes/templates/esyr/src/app.js
@@ -1,6 +1,6 @@
 const appComponentContent = ({ useTypeScript, pageName }) => `import React from 'react';
 import { createBrowserRouter, RouterProvider } from 'react-router-dom';
-import {${pageName}} from './pages/${pageName}';
+import {${pageName}} from './pages/${pageName}/${pageName}';
 
 const router = createBrowserRouter([
   {