+++ archetypes/templates/esyr/src/hooks.js
@@ -4,17 +4,20 @@ const { join } = require('path');
 /**
  * Builds a React hook file
  */
-const reactHookFactory = async ({ hookName, hookPath, useTypeScript }) => {
+const reactHookFactory = async ({ hookName, hookPath, useTypeScript, contentOverride }) => {
   const hookFileContent = `import React, { useEffect } from 'react';
 
 export const ${hookName} = () => {
 
+  ${contentOverride?.length > 0 ? contentOverride : `
   useEffect(() => {
     console.log('${hookName} mounted!');
   }, []);
 
-  return {}
-}`;
+  return {}`}
+}
+  
+export default ${hookName}`;
 
   const hookFileName = `${hookName}.${useTypeScript ? 'tsx' : 'jsx'}`;